### 2025-01-27 c122294
Provide kde option for density
