+++ huracanpy/calc/_density.py
@@ -4,9 +4,10 @@ Module containing function to compute track densities
 
 import numpy as np
 import xarray as xr
+from scipy.stats import gaussian_kde
 
 
-def density(lon, lat, *, method="histogram", bin_size=5, n_seasons=1):
+def density(lon, lat, method="histogram", bin_size=5, crop = False, function_kws = dict()):
     """Function to compute the track density, based on a simple 2D histogram.
 
     Parameters
@@ -35,29 +36,49 @@ def density(lon, lat, *, method="histogram", bin_size=5, n_seasons=1):
         Track density as a 2D map.
 
     """
-    if method == "histogram":
-        return _histogram(lon, lat, bin_size=bin_size) / n_seasons
-    else:
-        raise NotImplementedError(
-            f"Method {method} not implemented yet. Use one 'histogram'"
-        )
-
-
-def _histogram(lon, lat, bin_size=5):
+    # Define coordinates for mapping
     if lon.min() < 0:
         lon_range = (-180, 180)
     else:
         lon_range = (0, 360)
-    # compute 2D histogram
-    x = np.arange(lon_range[0], lon_range[1] + bin_size, bin_size)
-    y = np.arange(-90, 90 + bin_size, bin_size)
-    H, X, Y = np.histogram2d(lon, lat, bins=[x, y])
+    x_edge = np.arange(lon_range[0], lon_range[1] + bin_size, bin_size)
+    y_edge = np.arange(-90, 90 + bin_size, bin_size)
+    x_mid, y_mid = (x_edge[1:]+x_edge[:-1])/2, (y_edge[1:]+y_edge[:-1])/2
+    
+    # Compute density
+    if method == "histogram":
+        H =_histogram(lon, lat, x_edge, y_edge, function_kws)
+    elif method == "kde":
+        H = _kde(lon, lat, x_mid, y_mid, function_kws)
+    else:
+        raise NotImplementedError(
+            f"Method {method} not implemented yet. Use one 'histogram'"
+        )
+    
     # Turn into xarray
     da = xr.DataArray(
         H,
-        dims=["lon", "lat"],
-        coords={"lon": (X[:-1] + X[1:]) / 2, "lat": (Y[:-1] + Y[1:]) / 2},
+        dims=["lat", "lon"],
+        coords={"lon": x_mid, "lat": y_mid},
     )
-    # Format
-    da = da.where(da > 0).transpose()
-    return da.where(~np.isnan(da), drop=True).fillna(0)
+    
+    if crop: # Crop the map to where there are non-zero points
+        da = da.where(da > 0)
+        return da.where(~np.isnan(da), drop=True).fillna(0)
+    else: 
+        return da
+
+
+def _histogram(lon, lat, x_edge, y_edge, function_kws):
+    # Compute 2D histogram with numpy
+    H, _x, _y = np.histogram2d(lon, lat, bins=[x_edge, y_edge], **function_kws)
+    return H.T # Transpose result
+
+def _kde(lon, lat, x_mid, y_mid, function_kws):
+    # engineer positions array for kernel estimation computation
+    positions = np.reshape(np.meshgrid(x_mid, y_mid), (2, len(x_mid) * len(y_mid)))
+    # Compute kernel density estimate
+    kernel = gaussian_kde([lon, lat], **function_kws)
+    # Evaluation kernel along positions
+    return np.reshape(kernel(positions), (len(y_mid), len(x_mid)))
+    
