+++ .github/workflows/python-publish.yml
@@ -6,7 +6,7 @@
 # separate terms of service, privacy policy, and support
 # documentation.
 
-name: Upload Python Package
+name: Publish Python distribution to PyPI 
 
 on:
   release:
