+++ .github/workflows/python-publish.yml
@@ -11,6 +11,8 @@ name: Publish Python distribution to PyPI
 on:
   release:
     types: [published]
+  workflow_dispatch
+
 
 permissions:
   contents: read
