+++ huracanpy/_data/_load.py
@@ -116,6 +116,8 @@ def load(
           calendar to the default :class:`datetime.datetime`, then you can pass this
           argument to load the times in as :class:`cftime.datetime` with the given
           calendar instead
+          Can also be set to "timestep" if the TRACK file has timesteps instead of dates. 
+          In that case, you need to provide initial_date and timestep arguments. 
     **kwargs
         When loading tracks from a standard files these will be passed to the relevant
         load function
@@ -154,7 +156,7 @@ def load(
         source = source.lower()
         if source == "track":
             data = _TRACK.load(
-                filename, calendar=track_calendar, variable_names=variable_names
+                filename, calendar=track_calendar, variable_names=variable_names, **kwargs
             )
         elif source == "track.tilt":
             data = _TRACK.load_tilts(
