### 2025-02-06 9416917
Add error message when parsed line is empty because nb of tracks do not match header
