### 2025-02-06 a333124
Add new example file to inits
