### 2025-02-06 ef3fb79
Add test for new example file
