### 2025-02-10 f06eee4
Update github action to run checks for next version
