+++ tests/test_huracanpy.py
@@ -12,7 +12,18 @@ import huracanpy
         (huracanpy.example_csv_file, dict(), 9, 0, 99, 3),
         (huracanpy.example_parquet_file, dict(), 9, 0, 99, 3),
         (huracanpy.example_TRACK_netcdf_file, dict(), 20, 17, 4580, 86),
-        (huracanpy.example_TRACK_timestep_file, dict(source="TRACK", track_calendar="timestep", initial_date = "1940-01-01",), 38, 0, 416, 19),
+        (
+            huracanpy.example_TRACK_timestep_file,
+            dict(
+                source="TRACK",
+                track_calendar="timestep",
+                initial_date="1940-01-01",
+            ),
+            38,
+            0,
+            416,
+            19,
+        ),
         (huracanpy.example_TE_file, dict(source="tempestextremes"), 8, 0, 210, 8),
         (huracanpy.example_CHAZ_file, dict(), 11, 0, 1078, 20),
         (huracanpy.example_MIT_file, dict(), 10, 1, 2138, 11),
