+++ huracanpy/_data/_TRACK.py
@@ -1,6 +1,8 @@
 import gzip
 
 import datetime
+import warnings
+
 import cftime
 import numpy as np
 import xarray as xr
@@ -31,10 +33,10 @@ def _parse(fmt, string, **kwargs):
     return result
 
 
-def parse_date(date, calendar=None, initial_date = None, timestep=None):
+def parse_date(date, calendar=None, initial_date=None, timestep=None):
     if calendar == "timestep":
         initial_date = np.datetime64(initial_date)
-        delta = np.timedelta64((int(date)-1) * timestep, 'h') 
+        delta = np.timedelta64((int(date) - 1) * timestep, "h")
         return initial_date + delta
     elif len(date) == 10:  # i.e., YYYYMMDDHH
         if calendar is not None:
@@ -45,11 +47,11 @@ def parse_date(date, calendar=None, initial_date = None, timestep=None):
             return cftime.datetime(yr, mn, dy, hr, calendar=calendar)
         else:
             return datetime.datetime.strptime(date.strip(), "%Y%m%d%H")
-    else :
+    else:
         return int(date)
 
 
-def load(filename, calendar=None, initial_date = None, timestep = 6, variable_names=None):
+def load(filename, calendar=None, initial_date=None, timestep=6, variable_names=None):
     """Load ASCII TRACK data as an xarray.Dataset
 
     Parameters
@@ -131,31 +133,42 @@ def load(filename, calendar=None, initial_date = None, timestep = 6, variable_na
         for n in range(ntracks):
             # Read individual track header (two lines)
             line = f.readline().strip()
-            if not line.replace(' ', '') == '': # If line is empty
+            if not line.replace(" ", "") == "":  # If line is empty
                 try:
                     track_info = _parse(track_header_fmt, line).named
                 except ValueError:
                     track_info = _parse(track_header_fmt_new, line).named
-    
+
                 line = f.readline().strip()
                 npoints = _parse(track_info_fmt, line)["npoints"]
-    
+
                 # Generate arrays for time coordinate and variables
                 # Time is a list because it will hold datetime or cftime objects
                 # Other variables are a dictionary mapping variable name to a tuple of
                 # (time, data_array) as this is what is passed to xarray.Dataset
-    
+
                 # Add track ID as a variable along the record dimension so that it can be used
                 # for groupby
                 track_id.extend([track_info["track_id"]] * npoints)
-    
+
                 # Populate time and data line by line
                 for m in range(npoints):
                     line = f.readline().strip()
-                    times.append(parse_date(line.split(" ")[0], calendar=calendar, initial_date=initial_date, timestep=timestep))
+                    times.append(
+                        parse_date(
+                            line.split(" ")[0],
+                            calendar=calendar,
+                            initial_date=initial_date,
+                            timestep=timestep,
+                        )
+                    )
                     output.append(line.replace("&", " "))
-            else: 
-                print("Parsed line is empty. It is possible this problem arrises because the number of tracks expected from the header was not found in the file.") # TODO: Make a proper error message.
+            else:
+                warnings.warn(
+                    "Parsed line is empty. It is possible this problem arrises because"
+                    "the number of tracks expected from the header was not found in the"
+                    "file."
+                )
 
     output = np.genfromtxt(output)
     data_vars = {
