### 2025-02-12 92078bc
Create functions to read 'old HURDAT'/ECMWF files
