### 2025-02-12 2a5fa1c
Register new 'old HURDAT'/ECMWF example file
