+++ huracanpy/__init__.py
@@ -26,6 +26,7 @@ __all__ = [
     "example_MIT_file",
     "example_ERA20C_file",
     "example_WiTRACK_file",
+    "example_old_HURDAT_file",
     "_accessor",
 ]
 
@@ -43,6 +44,7 @@ from ._data import (
     example_MIT_file,
     example_ERA20C_file,
     example_WiTRACK_file,
+    example_old_HURDAT_file,
 )
 from ._basins import basins
 from ._interp import interp_time
