### 2025-02-12 51f3ea4
Add option to load 'old HURDAT'/ECMWF tracks
