### 2025-02-12 464488c
Don't run ipynb's with tests by default
