+++ tests/test_docs.py
@@ -13,6 +13,10 @@ all_notebooks = [(str(fname),) for fname in docs_path.glob("examples/*.ipynb")] 
 ]
 
 
+@pytest.mark.skipif(
+    "not config.getoption('--docs')",
+    reason="ipynb's are slow to run. Specify --docs explicitly to test execution",
+)
 @pytest.mark.parametrize(("notebook",), all_notebooks)
 def test_notebook_runs(notebook):
     with open(notebook) as f:
