### 2025-02-12 89a5e7a
Add test for 'old HURDAT'/ECMWF
