### 2025-02-13 4c9ace2
Add function to compute azimuth
