### 2025-02-13 be30c87
Add test for azimuth
