### 2025-02-13 7379520
Add accessor azimuth functions test
