+++ huracanpy/_data/__init__.py
@@ -54,4 +54,4 @@ example_MIT_file = str(testdata_dir / "MIT_sample.nc")
 example_ERA20C_file = str(testdata_dir / "ERA20C_TRACK_NATL_tcident.nc")
 example_WiTRACK_file = str(testdata_dir / "event_tracks_2017_ERA5.dat")
 
-example_old_HURDAT_file =  str(testdata_dir / "00012024050100_050_0005160_atl")
+example_old_HURDAT_file = str(testdata_dir / "00012024050100_050_0005160_atl")
