### 2025-02-18 b7898ba
Update requirements for docs test
