### 2025-02-24 b5f022d
Add test for sel_id
