### 2025-03-21 #7670 @hugoledoux
@stella-bourdin could you please while we look for reviewers perhaps improve the `README.md`? Now it's just one line and then an advertisement... I would copy some relevant info from the docs.

Your docs is failing also, the badge says?
