### 2025-04-03 a3485e6
Update README.md with snippets from the user guide
