### 2025-04-03 #7670 @stella-bourdin
> [@stella-bourdin](https://github.com/stella-bourdin) could you please while we look for reviewers perhaps improve the `README.md`? Now it's just one line and then an advertisement... I would copy some relevant info from the docs.
>
> Your docs is failing also, the badge says?

Hi,
Thank you for editing our paper!
Is it possible to add @leosaffin to pings? We are co-developping the package.
Leo has just updated the README and fixed the doc.
