### 2025-04-07 9b29a47
Add infer_track_id function and include as an option to load. Load in a STORM file with these options as a test
