+++ docs/api/huracanpy.rst
@@ -12,3 +12,4 @@ huracanpy
     sel_id
     trackswhere
     interp_time
+    infer_track_id
