### 2025-04-08 9d13c8a
Add equivalent functions to accessor and move to utils.inferred_track_id for naming to make better sense
