### 2025-04-08 21035ab
Add equivalent functions to accessor and move to utils.inferred_track_id for naming to make better sense
