+++ huracanpy/_data/_load.py
@@ -3,7 +3,7 @@ import pandas as pd
 
 from . import _csv, _TRACK, _netcdf, _tempestextremes, witrack, _old_HURDAT
 from . import ibtracs
-from .. import _utils
+from .. import info
 
 
 rename_defaults = dict(
@@ -223,7 +223,7 @@ def load(
     if infer_track_id is not None:
         data["track_id"] = (
             "record",
-            _utils.infer_track_id(*[data[var] for var in infer_track_id]),
+            info.inferred_track_id(*[data[var] for var in infer_track_id]),
         )
 
     return data
