+++ docs/api/info.rst
@@ -35,3 +35,11 @@ Time
 
     season
     time_components
+
+Other
+-----
+
+.. autosummary::
+    :toctree: _autosummary
+
+    inferred_track_id
