+++ docs/api/huracanpy.rst
@@ -12,4 +12,3 @@ huracanpy
     sel_id
     trackswhere
     interp_time
-    infer_track_id
