### 2025-04-08 3662bcd
Try to return the same type as the input to inferred_track_id. Makes the accessor easier
