### 2025-04-08 d9f450c
Try to return the same type as the input to inferred_track_id. Makes the accessor easier
