+++ tests/test_info/test_utils.py
@@ -1,4 +1,4 @@
-import numpy as np
+import xarray as xr
 
 import huracanpy
 
@@ -7,4 +7,4 @@ def test_inferred_track_id():
     tracks = huracanpy.load(huracanpy.example_csv_file)
     track_id = huracanpy.info.inferred_track_id(tracks.track_id)
 
-    np.testing.assert_equal(track_id, tracks.track_id)
+    xr.testing.assert_equal(track_id, tracks.track_id)
