+++ .github/workflows/python-package.yaml
@@ -5,9 +5,9 @@ name: Python package
 
 on:
   push:
-    branches: [ "main", "v1.2" ]
+    branches: [ "master", "v1.*" ]
   pull_request:
-    branches: [ "main", "v1.2" ]
+    branches: [ "master", "v1.*" ]
 
 jobs:
   build:
