### 2025-04-09 9e8b940
Add conda install instructions
