+++ docs/user_guide/install.md
@@ -1,0 +1,11 @@
+# Installation
+
+You can install huracanpy with conda
+```bash
+conda install -c conda-forge huracanpy
+```
+or pip
+```bash
+pip install huracanpy
+```
+For older python versions, the pip installation may fail due to issues with installing cartopy through pip. If this happens, use conda to install cartopy first (e.g. ``conda install -c conda-forge cartopy``), then install huracanpy.
