### 2025-04-09 85190dd
Add conda install instructions
