+++ docs/user_guide/install.rst
@@ -1,11 +1,0 @@
-Installation
-============
-
-To install the package, you can use ``pip``::
-
-    pip install huracanpy
-
-Nota Bene:
-
-#. We support Python 3.9 to 3.13.
-#. The installation may fail due to issues with installing cartopy through pip. If this happens, use conda to install cartopy first (e.g. ``conda install -c conda-forge cartopy``), then install huracanpy as normal.
