### 2025-04-16 76dc8d3
Add hrcn accessor methods to API
