+++ pyproject.toml
@@ -46,6 +46,7 @@ docs = [
     "sphinx-design",
     "ipykernel",
     "myst-parser",
+    "sphinx_autosummary_accessors",
 ]
 
 [build-system]
