### 2025-05-07 #8174 @editorialbot
[REVIEW]: HuracanPy: A python package for reading and analysing cyclone tracks
**Submitting author:** @stella-bourdin (<a href="http://orcid.org/0000-0003-2635-5654">Stella Bourdin</a>)
**Repository:** https://github.com/Huracan-project/huracanpy
**Branch with paper.md** (empty if default branch): JOSS-paper
**Version:** v1.3.0
