### 2025-05-12 Huracan-project/huracanpy#103 @nmstreethran
[JOSS Review] Tests
Opening this issue as part of my JOSS review (https://github.com/openjournals/joss-reviews/issues/8174). This issue concerns the [Tests](https://huracanpy.readthedocs.io/en/latest/dev_guide/tests.html) page of the documentation.

### Running tests

The following text is under "Running tests", in the second paragraph. If you're not sure then you should remove that part from the docs. ;)

> I used conda environments and symlinks to achieve this. I don’t know if that’s how you’re supposed to do it.

### Pytest results

I ran `pytest` and there were a significant number of skips and warnings. This includes several deprecation warnings which should be fixed.

Can you review the log below and fix these warnings?

<details><summary>Details</summary>
<pre>
============================================================================= test session starts =============================================================================
rootdir: /mnt/Backup/Downloads/huracanpy
configfile: pyproject.toml
collected 208 items

tests/test_accel.py ...............                                                                                                                                     [  7%]
tests/test_accessor.py ................s...................s....                                                                                                        [ 26%]
tests/test_assess.py ..                                                                                                                                                 [ 27%]
tests/test_calc/test_lifecycle.py ..                                                                                                                                    [ 28%]
tests/test_calc/test_rates.py .....................                                                                                                                     [ 38%]
tests/test_calc/test_track_density.py .                                                                                                                                 [ 39%]
tests/test_calc/test_track_stats.py ...                                                                                                                                 [ 40%]
tests/test_calc/test_translation.py ....                                                                                                                                [ 42%]
tests/test_docs.py sssssssssssssssssss                                                                                                                                  [ 51%]
tests/test_huracanpy.py ...............ss...ss.............ss...ss............                                                                                          [ 77%]
tests/test_info/test_category.py ..                                                                                                                                     [ 78%]
tests/test_info/test_geography.py .................                                                                                                                     [ 87%]
tests/test_info/test_time.py ...                                                                                                                                        [ 88%]
tests/test_interpolate.py .                                                                                                                                             [ 88%]
tests/test_subset.py ...                                                                                                                                                [ 90%]
tests/test_tc/test_ace.py ...                                                                                                                                           [ 91%]
tests/test_tc/test_beta_drift.py .                                                                                                                                      [ 92%]
tests/test_tc/test_tc_category.py ................                                                                                                                      [100%]

============================================================================== warnings summary ===============================================================================
.venv/lib/python3.12/site-packages/geopandas/_compat.py:7
  /mnt/Backup/Downloads/huracanpy/.venv/lib/python3.12/site-packages/geopandas/_compat.py:7: DeprecationWarning: The 'shapely.geos' module is deprecated, and will be removed in a future version. All attributes of 'shapely.geos' are available directly from the top-level 'shapely' namespace (since shapely 2.0.0).
    import shapely.geos

.venv/lib/python3.12/site-packages/jupyter_client/connect.py:22
  /mnt/Backup/Downloads/huracanpy/.venv/lib/python3.12/site-packages/jupyter_client/connect.py:22: DeprecationWarning: Jupyter is migrating its paths to use standard platformdirs
  given by the platformdirs library.  To remove this warning and
  see the appropriate new directories, set the environment variable
  `JUPYTER_PLATFORM_DIRS=1` and then run `jupyter --paths`.
  The use of platformdirs will be the default in `jupyter_core` v6
    from jupyter_core.paths import jupyter_data_dir, jupyter_runtime_dir, secure_write

tests/test_accel.py::test_accel_get_gen_vals[tracks_csv]
tests/test_accel.py::test_accel_get_gen_vals[tracks_with_extra_coord]
tests/test_accel.py::test_accel_get_apex_vals[tracks_csv]
tests/test_accel.py::test_accel_get_apex_vals[tracks_with_extra_coord]
  /mnt/Backup/Downloads/huracanpy/.venv/lib/python3.12/site-packages/xarray/structure/concat.py:540: UserWarning: No index created for dimension track_id because variable track_id is not a coordinate. To create an index for track_id, please first call `.set_coords('track_id')` on this object.
    ds.expand_dims(dim_name, create_index_for_new_dim=create_index_for_new_dim)

tests/test_accel.py::test_accel_get_time_from_genesis[tracks_csv]
tests/test_accel.py::test_accel_get_time_from_genesis[tracks_with_extra_coord]
tests/test_accessor.py::test_accessor_methods_match_functions[time_from_genesis-function_args15-time_from_genesis-accessor_function_kwargs15-get]
tests/test_accessor.py::test_accessor_methods_match_functions[time_from_genesis-function_args15-time_from_genesis-accessor_function_kwargs15-get]
tests/test_accessor.py::test_accessor_methods_match_functions[time_from_genesis-function_args15-time_from_genesis-accessor_function_kwargs15-add]
tests/test_accessor.py::test_accessor_methods_match_functions[time_from_genesis-function_args15-time_from_genesis-accessor_function_kwargs15-add]
tests/test_calc/test_lifecycle.py::test_time_from_genesis
  /mnt/Backup/Downloads/huracanpy/huracanpy/calc/_lifecycle.py:36: DeprecationWarning: dropping variables using `drop` is deprecated; use drop_vars.
    .drop(track_ids.dims[0])

tests/test_accel.py::test_accel_get_time_from_apex[tracks_csv]
tests/test_accel.py::test_accel_get_time_from_apex[tracks_with_extra_coord]
tests/test_accessor.py::test_accessor_methods_match_functions[time_from_apex-function_args16-time_from_apex-accessor_function_kwargs16-get]
tests/test_accessor.py::test_accessor_methods_match_functions[time_from_apex-function_args16-time_from_apex-accessor_function_kwargs16-get]
tests/test_accessor.py::test_accessor_methods_match_functions[time_from_apex-function_args16-time_from_apex-accessor_function_kwargs16-add]
tests/test_accessor.py::test_accessor_methods_match_functions[time_from_apex-function_args16-time_from_apex-accessor_function_kwargs16-add]
tests/test_calc/test_lifecycle.py::test_time_from_extremum
  /mnt/Backup/Downloads/huracanpy/huracanpy/calc/_lifecycle.py:76: DeprecationWarning: dropping variables using `drop` is deprecated; use drop_vars.
    .drop(track_ids.dims[0])

tests/test_accel.py::test_accel_get_time_from_apex[tracks_csv]
tests/test_accel.py::test_accel_get_time_from_apex[tracks_csv]
tests/test_accel.py::test_accel_get_time_from_apex[tracks_csv]
tests/test_accel.py::test_accel_get_time_from_apex[tracks_with_extra_coord]
tests/test_accel.py::test_accel_get_time_from_apex[tracks_with_extra_coord]
tests/test_accel.py::test_accel_get_time_from_apex[tracks_with_extra_coord]
  /mnt/Backup/Downloads/huracanpy/.venv/lib/python3.12/site-packages/xarray/core/dataarray.py:6313: DeprecationWarning: Behaviour of argmin/argmax with neither dim nor axis argument will change to return a dict of indices of each dimension. To get a single, flat index, please use np.argmin(da.data) or np.argmax(da.data) instead of da.argmin() or da.argmax().
    result = self.variable.argmax(dim, axis, keep_attrs, skipna)

tests/test_accessor.py: 20 warnings
tests/test_info/test_geography.py: 15 warnings
  /mnt/Backup/Downloads/huracanpy/huracanpy/info/_geography.py:116: DeprecationWarning: __array__ implementation doesn't accept a copy keyword, so passing copy=False failed. __array__ must implement 'dtype' and 'copy' keyword arguments. To learn more, see the migration guide https://numpy.org/devdocs/numpy_2_0_migration_guide.html#adapting-to-changes-in-the-copy-keyword
    lon = np.array(lon)

tests/test_accessor.py: 20 warnings
tests/test_info/test_geography.py: 15 warnings
  /mnt/Backup/Downloads/huracanpy/huracanpy/info/_geography.py:117: DeprecationWarning: __array__ implementation doesn't accept a copy keyword, so passing copy=False failed. __array__ must implement 'dtype' and 'copy' keyword arguments. To learn more, see the migration guide https://numpy.org/devdocs/numpy_2_0_migration_guide.html#adapting-to-changes-in-the-copy-keyword
    lat = np.array(lat)

tests/test_accessor.py::test_accessor_methods_match_functions[is_land-function_args2-is_land-accessor_function_kwargs2-get]
  /mnt/Backup/Downloads/huracanpy/.venv/lib/python3.12/site-packages/cartopy/io/__init__.py:241: DownloadWarning: Downloading: https://naturalearth.s3.amazonaws.com/10m_physical/ne_10m_ocean.zip
    warnings.warn(f'Downloading: {url}', DownloadWarning)

tests/test_accessor.py::test_accessor_methods_match_functions[country-function_args4-country-accessor_function_kwargs4-get]
  /mnt/Backup/Downloads/huracanpy/.venv/lib/python3.12/site-packages/cartopy/io/__init__.py:241: DownloadWarning: Downloading: https://naturalearth.s3.amazonaws.com/10m_cultural/ne_10m_admin_0_countries.zip
    warnings.warn(f'Downloading: {url}', DownloadWarning)

tests/test_accessor.py::test_accessor_methods_match_functions[season-function_args8-season-accessor_function_kwargs8-get]
tests/test_accessor.py::test_accessor_methods_match_functions[season-function_args8-season-accessor_function_kwargs8-get]
tests/test_accessor.py::test_accessor_methods_match_functions[season-function_args8-season-accessor_function_kwargs8-add]
tests/test_accessor.py::test_accessor_methods_match_functions[season-function_args8-season-accessor_function_kwargs8-add]
  /mnt/Backup/Downloads/huracanpy/.venv/lib/python3.12/site-packages/pandas/core/arrays/datetimelike.py:2452: UnitStrippedWarning: The unit of the quantity is stripped when downcasting to ndarray.
    data = np.asarray(data)

tests/test_accessor.py::test_accessor_methods_match_functions[season-function_args8-season-accessor_function_kwargs8-get]
tests/test_accessor.py::test_accessor_methods_match_functions[season-function_args8-season-accessor_function_kwargs8-get]
tests/test_accessor.py::test_accessor_methods_match_functions[season-function_args8-season-accessor_function_kwargs8-add]
tests/test_accessor.py::test_accessor_methods_match_functions[season-function_args8-season-accessor_function_kwargs8-add]
tests/test_info/test_time.py::test_seasons[tracks_year]
tests/test_info/test_time.py::test_seasons[tracks_year_cftime]
  /mnt/Backup/Downloads/huracanpy/.venv/lib/python3.12/site-packages/pandas/core/construction.py:630: UnitStrippedWarning: The unit of the quantity is stripped when downcasting to ndarray.
    data = np.asarray(data)

tests/test_accessor.py: 4 warnings
tests/test_info/test_category.py: 2 warnings
tests/test_subset.py: 1 warning
tests/test_tc/test_tc_category.py: 17 warnings
  /mnt/Backup/Downloads/huracanpy/.venv/lib/python3.12/site-packages/pandas/core/indexes/base.py:528: UnitStrippedWarning: The unit of the quantity is stripped when downcasting to ndarray.
    return cls(np.asarray(data), dtype=dtype, copy=copy, name=name)

tests/test_accessor.py::test_accessor_methods_match_functions[pressure_category-function_args10-pressure_category-accessor_function_kwargs10-get]
tests/test_accessor.py::test_accessor_methods_match_functions[pressure_category-function_args10-pressure_category-accessor_function_kwargs10-add]
tests/test_tc/test_tc_category.py::test_pressure_cat_units[Pa-default-Klotzbach-False]
tests/test_tc/test_tc_category.py::test_pressure_cat_units[Pa-default-Klotzbach-True]
tests/test_tc/test_tc_category.py::test_pressure_cat_units[Pa-default-Simpson-False]
tests/test_tc/test_tc_category.py::test_pressure_cat_units[Pa-default-Simpson-True]
tests/test_tc/test_tc_category.py::test_pressure_cat_units[hPa-expected1-Klotzbach-True]
tests/test_tc/test_tc_category.py::test_pressure_cat_units[hPa-expected1-Simpson-True]
  /mnt/Backup/Downloads/huracanpy/huracanpy/tc/_category.py:69: UserWarning: Caution, pressure are likely in Pa, they are being converted to hPa for categorization. In future specify the units explicitly by passing slp_units="Pa" to this function or setting slp.attrs["units"] = "Pa"
    warnings.warn(

tests/test_accessor.py::test_accessor_methods_match_functions[distance-function_args11-distance-accessor_function_kwargs11-get]
tests/test_accessor.py::test_accessor_methods_match_functions[distance-function_args11-distance-accessor_function_kwargs11-add]
tests/test_accessor.py::test_accessor_methods_match_functions[azimuth-function_args12-azimuth-accessor_function_kwargs12-get]
tests/test_accessor.py::test_accessor_methods_match_functions[azimuth-function_args12-azimuth-accessor_function_kwargs12-add]
tests/test_accessor.py::test_accessor_methods_match_functions[translation_speed-function_args13-translation_speed-accessor_function_kwargs13-get]
tests/test_accessor.py::test_accessor_methods_match_functions[translation_speed-function_args13-translation_speed-accessor_function_kwargs13-add]
tests/test_calc/test_translation.py::test_azimuth
tests/test_calc/test_translation.py::test_get_distance
tests/test_calc/test_translation.py::test_get_translation_speed
  /mnt/Backup/Downloads/huracanpy/.venv/lib/python3.12/site-packages/pyproj/utils.py:121: UnitStrippedWarning: The unit of the quantity is stripped when downcasting to ndarray.
    xxx = xxx.__array__()

tests/test_accessor.py::test_get_methods
tests/test_accessor.py::test_get_methods
tests/test_tc/test_ace.py::test_pace
tests/test_tc/test_ace.py::test_pace
  /mnt/Backup/Downloads/huracanpy/.venv/lib/python3.12/site-packages/numpy/polynomial/polyutils.py:353: UnitStrippedWarning: The unit of the quantity is stripped when downcasting to ndarray.
    x = np.asanyarray(x)

tests/test_assess.py::test_match
tests/test_assess.py::test_scores
tests/test_huracanpy.py::test_load[None-kwargs11-8-0-143287-4540]
tests/test_huracanpy.py::test_load[None-kwargs12-10-0-121806-4170]
tests/test_huracanpy.py::test_save[False-csv-None-ibtracs]
tests/test_huracanpy.py::test_save[False-nc-None-ibtracs]
tests/test_huracanpy.py::test_save[True-csv-None-ibtracs]
tests/test_huracanpy.py::test_save[True-nc-None-ibtracs]
  /mnt/Backup/Downloads/huracanpy/huracanpy/_data/ibtracs.py:112: UserWarning: This offline function loads a light version of IBTrACS which is embedded within the package, based on a file produced manually by the developers.
                    It was last updated on the 15th Nov 2024, based on the IBTrACS file at that date.
                    It contains only data from 1980 up to the last year with no provisional tracks. All spur tracks were removed. Only 6-hourly time steps were kept.
    warnings.warn(

tests/test_assess.py::test_match
tests/test_assess.py::test_scores
tests/test_huracanpy.py::test_load[None-kwargs11-8-0-143287-4540]
tests/test_huracanpy.py::test_save[False-csv-None-ibtracs]
tests/test_huracanpy.py::test_save[False-nc-None-ibtracs]
tests/test_huracanpy.py::test_save[True-csv-None-ibtracs]
tests/test_huracanpy.py::test_save[True-nc-None-ibtracs]
  /mnt/Backup/Downloads/huracanpy/huracanpy/_data/ibtracs.py:118: UserWarning: You are loading the IBTrACS-WMO subset.                       This dataset contains the positions and intensity reported by the WMO agency responsible for each basin
                        Be aware of the fact that wind and pressure data is provided as they are in IBTrACS,                       which means in particular that wind speeds are in knots and averaged over different time periods.
                      For more information, see the IBTrACS column documentation at https://www.ncei.noaa.gov/sites/default/files/2021-07/IBTrACS_v04_column_documentation.pdf
    warnings.warn(

tests/test_calc/test_rates.py: 12 warnings
  /mnt/Backup/Downloads/huracanpy/huracanpy/calc/_rates.py:38: UserWarning: track_id is not provided, all points are considered to come from the sametrack
    warnings.warn(

tests/test_calc/test_translation.py::test_get_distance
tests/test_calc/test_translation.py::test_radius_of_maximum_wind
tests/test_tc/test_beta_drift.py::test_radius_of_maximum_wind
  /mnt/Backup/Downloads/huracanpy/huracanpy/calc/_translation.py:33: UnitStrippedWarning: The unit of the quantity is stripped when downcasting to ndarray.
    yx1 = np.array([lat1, lon1]).T

tests/test_calc/test_translation.py::test_get_distance
tests/test_calc/test_translation.py::test_radius_of_maximum_wind
tests/test_tc/test_beta_drift.py::test_radius_of_maximum_wind
  /mnt/Backup/Downloads/huracanpy/huracanpy/calc/_translation.py:34: UnitStrippedWarning: The unit of the quantity is stripped when downcasting to ndarray.
    yx2 = np.array([lat2, lon2]).T

tests/test_huracanpy.py::test_load[/mnt/Backup/Downloads/huracanpy/huracanpy/_data/example_data/tr_trs_pos.2day_addT63vor_addmslp_add925wind_add10mwind.tcident.new.nc-kwargs4-20-17-4580-86]
  <frozen importlib._bootstrap>:488: RuntimeWarning: numpy.ndarray size changed, may indicate binary incompatibility. Expected 16 from C header, got 96 from PyObject

tests/test_info/test_time.py::test_seasons[tracks_year]
tests/test_info/test_time.py::test_seasons[tracks_year_cftime]
  /mnt/Backup/Downloads/huracanpy/.venv/lib/python3.12/site-packages/pandas/core/tools/datetimes.py:243: DeprecationWarning: __array__ implementation doesn't accept a copy keyword, so passing copy=False failed. __array__ must implement 'dtype' and 'copy' keyword arguments. To learn more, see the migration guide https://numpy.org/devdocs/numpy_2_0_migration_guide.html#adapting-to-changes-in-the-copy-keyword
    arg = np.array(arg)

tests/test_info/test_time.py::test_seasons[tracks_year]
tests/test_info/test_time.py::test_seasons[tracks_year_cftime]
  /mnt/Backup/Downloads/huracanpy/.venv/lib/python3.12/site-packages/pandas/core/tools/datetimes.py:243: UnitStrippedWarning: The unit of the quantity is stripped when downcasting to ndarray.
    arg = np.array(arg)

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
========================================================== 179 passed, 29 skipped, 191 warnings in 62.90s (0:01:02) ===========================================================
</pre>

### Writing tests

The following is under "Writing tests", but it would be more useful to give a brief description similar to the other assert methods:

> assert - The old classic
