### 2025-05-12 Huracan-project/huracanpy#104 @nmstreethran
[JOSS Review] Installation
Opening this issue as part of my JOSS review (https://github.com/openjournals/joss-reviews/issues/8174). This issue concerns the installation instructions in the documentation.

### User guide > Installation

I recommend including steps to create a virtual environment in the installation command and prefixing the pip command with `python -m`.

```sh
python -m venv .venv
source .venv/bin/activate
python -m pip install huracanpy
```

Additionally, I suggest including instructions for installing from source, for those wanting to use the most up-to-date package as available on GitHub. See [here](https://pip.pypa.io/en/stable/topics/vcs-support/) for more information.

```sh
python -m venv .venv
source .venv/bin/activate
python -m pip install "huracanpy @ git+https://github.com/Huracan-project/huracanpy"
```

Under "Nota Bene", the following is stated. I'm not aware of any issues installing cartopy through pip. Can you elaborate? I don't think installing cartopy using conda is the best option here, as there will be a mixture of conda and pip packages, and no guidelines are provided for setting up conda and virtual environments.

> The installation may fail due to issues with installing cartopy through pip. If this happens, use conda to install cartopy first (e.g. conda install -c conda-forge cartopy), then install huracanpy as normal.

### Developer installation

- Remove the indentations at the start of the commands.

- This command causes an error. Remove the slash before `https` to fix it:

  ```sh
      git clone /https://github.com/{your-username}/huracanpy.git
  ```

  ```
  Cloning into 'huracanpy'...
  fatal: protocol '/https' is not supported
  ```

- I suggest including the following steps for completeness:

    - Navigate to the directory of the cloned repository:

      ```sh
      cd huracanpy
      ```

    - Create a virtual environment before installing any packages, which is recommended:

      ```sh
      python -m venv .venv
      source .venv/bin/activate
      ```

- The pip installation command causes the following error. Remove the space after the comma to fix it.

  ```sh
  python -m pip install -e .[dev, docs]
  ```

  ```
  ERROR: Invalid requirement: 'docs]': Expected end or semicolon (after name and no valid version specifier)
      docs]
          ^
  ```

Under "2. ruff", the link is incorrectly formatted:

> We use [ruff](<https://docs.astral.sh/ruff/) to automatically keep the style of the code consistent, so we don’t have to worry about it.
