+++ tests/test_huracanpy.py
@@ -72,7 +72,7 @@ import huracanpy
     ],
 )
 def test_load(filename, kwargs, nvars, ncoords, npoints, ntracks):
-    data = huracanpy.load(filename, **kwargs)
+    data = _load_with_checked_warnings(filename, **kwargs)
 
     assert len(data) == nvars
     assert len(data.coords) == ncoords
@@ -112,7 +112,7 @@ def test_save(filename, source, extension, muddle, tmp_path):
             " dataframe leads to having rows equal to the product of the dimensions"
             " even though the dimensions cover different variables"
         )
-    data = huracanpy.load(filename, source=source)
+    data = _load_with_checked_warnings(filename, source=source)
 
     # Check that save/load gives the same result when the track_id is not monotonic
     # Caused an issue because they got sorted before
@@ -132,6 +132,31 @@ def test_save(filename, source, extension, muddle, tmp_path):
         data = data.sortby("track_id")
     data_reload = huracanpy.load(str(tmp_path / f"tmp_file.{extension}"))
     _assert_dataset_identical(data, data_reload)
+
+
+def _load_with_checked_warnings(filename, **kwargs):
+    if filename is None:
+        if "ibtracs_subset" not in kwargs or kwargs["ibtracs_subset"] == "wmo":
+            with (
+                pytest.warns(
+                    UserWarning,
+                    match="This offline function loads a light version of IBTrACS",
+                ),
+                pytest.warns(
+                    UserWarning, match="You are loading the IBTrACS-WMO subset"
+                ),
+            ):
+                data = huracanpy.load(filename, **kwargs)
+        else:
+            with pytest.warns(
+                UserWarning,
+                match="This offline function loads a light version of IBTrACS",
+            ):
+                data = huracanpy.load(filename, **kwargs)
+    else:
+        data = huracanpy.load(filename, **kwargs)
+
+    return data
 
 
 def _assert_dataset_identical(ds1, ds2):
