+++ tests/test_accessor.py
@@ -43,7 +43,7 @@ def test_nunique():
             huracanpy.tc.pressure_category,
             ["slp"],
             "pressure_category",
-            {"slp_name": "slp"},
+            {"slp_name": "slp", "slp_units": "Pa"},
         ),
         (huracanpy.calc.distance, ["lon", "lat", "track_id"], "distance", {}),
         (huracanpy.calc.azimuth, ["lon", "lat", "track_id"], "azimuth", {}),
@@ -120,7 +120,11 @@ def test_accessor_methods_match_functions(
     function_args = [
         tracks_csv[var] if not var == "all" else tracks_csv for var in function_args
     ]
-    result = function(*function_args)
+    if function == huracanpy.tc.pressure_category:
+        with pytest.warns(UserWarning, match="Caution, pressure are likely in Pa"):
+            result = function(*function_args)
+    else:
+        result = function(*function_args)
 
     # Call the accessor method
     result_accessor = getattr(tracks_csv.hrcn, f"{call_type}_{accessor_name}")(
