### 2025-05-19 fb99e58
Fix remaining deprecation warnings
