### 2025-05-19 12bdd6c
Fix remaining deprecation warnings
