+++ tests/conftest.py
@@ -7,8 +7,7 @@ import cftime
 import numpy as np
 import xarray as xr
 
-with pytest.warns(PendingDeprecationWarning, match="Overriding `Axes.cla` in GeoAxes"):
-    import huracanpy
+import huracanpy
 
 
 def pytest_addoption(parser):
