### 2025-05-19 0eb621d
Ignore cartopy download warnings in tests
