### 2025-05-19 Huracan-project/huracanpy#103 @leosaffin
Thanks for the suggestion. The pull request (#105) should have fixed all of these. The tests have been changed to give an error for any warnings and then fixed by either suppressing the warning, asserting that the warning is raised, or ignoring the warning (in that order of priority).
