### 2025-05-19 1c3fccb
Update install instructions
