### 2025-05-19 643a8d8
Update install instructions
