### 2025-05-19 Huracan-project/huracanpy#104 @leosaffin
Thanks for the comments. See #106 for full changes. For the following changes, I have done exactly what you suggested
- Add `python -m` to pip commands
- Add instructions to install from source
- Remove Nota Bene about cartopy. This was an issue with older versions of python/cartopy, but doesn't seem to be a problem anymore
- Remove indentations
- Fix git clone command
- Add `cd` to install instructions
- Fix dev,docs command
- Fix ruff link

The only thing I have done differently is the instructions for setting up a virtual environment. Huracanpy is intended as a python library, not a standalone tool, so I would expect most users would install it in an existing environment (or already have created a new environment. In the developer instructions I have put commands for setting up a fresh environment, with conda or venv. I'm essentially copying what xarray have done with their documentation for [users](https://docs.xarray.dev/en/stable/getting-started-guide/installing.html#instructions) and [developers](https://docs.xarray.dev/en/stable/contribute/contributing.html#contributing-dev-env).

Huracanpy is also now on conda-forge, so that has been added to the instructions with conda shown as the first option
