+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={8},
   pages={5023--5048},
   year={2021},
-  publisher={Copernicus Publications G{\"o}ttingen, Germany}
+  publisher={Copernicus Publications G{\"o}ttingen, Germany},
+  doi={10.5194/gmd-14-5023-2021}
 }
 
 
@@ -122,7 +123,8 @@ Publisher: American Meteorological Society},
   number={6},
   pages={E830--E846},
   year={2020},
-  publisher={American Meteorological Society}
+  publisher={American Meteorological Society},
+  doi={10.1175/bams-d-19-0062.1},
 }
 
 @article{befort2020objective,
@@ -133,7 +135,8 @@ Publisher: American Meteorological Society},
   number={3},
   pages={031005},
   year={2020},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/2515-7620/ab7b35},
 }
 
 @misc{gahtan2024international,
