+++ huracanpy/_data/__init__.py
@@ -17,6 +17,7 @@ __all__ = [
     "example_WiTRACK_file",
     "example_old_HURDAT_file",
     "example_STORM_file",
+    "example_IRIS_file",
 ]
 
 import pathlib
@@ -58,3 +59,5 @@ example_WiTRACK_file = str(testdata_dir / "event_tracks_2017_ERA5.dat")
 example_old_HURDAT_file = str(testdata_dir / "00012024050100_050_0005160_atl")
 
 example_STORM_file = str(testdata_dir / "STORM_DATA_NA_10_YEARS_0.txt")
+
+example_IRIS_file = str(testdata_dir / "IRIS_NI_1000Y_n6.txt")
