+++ huracanpy/_data/example_data/IRIS_NI_1000Y_n6.txt
@@ -1,0 +1,41 @@
+#tcid basin year tc month timestep lon lat vmax pmin rmw r18
+332829 NI 6000 0 10 0 71.7238 17.5157 76.01 894.34 27.8 378.3
+332829 NI 6000 0 10 1 71.6927 18.1946 72.6 901.23 29.5 361.8
+332829 NI 6000 0 10 2 71.6541 18.8134 69.48 907.42 31.1 345.3
+332829 NI 6000 0 10 3 71.6367 19.216 66.61 913.07 32.7 328.8
+332829 NI 6000 0 10 4 71.6344 19.5111 63.98 918.21 34.4 312.4
+332829 NI 6000 0 10 5 71.5259 20.0505 61.54 922.88 36.0 295.9
+332829 NI 6000 0 10 6 71.4147 20.6088 59.29 927.16 37.7 279.4
+332829 NI 6000 0 10 7 71.4322 20.8914 57.19 931.16 39.3 262.9
+332829 NI 6000 0 10 8 71.495 21.1066 55.23 934.86 40.9 246.4
+332829 NI 6000 0 10 9 71.5389 21.3991 42.84 954.99 42.6 229.9
+332829 NI 6000 0 10 10 71.5753 21.7543 34.98 966.67 44.2 213.4
+332829 NI 6000 0 10 11 71.6017 22.1856 29.56 974.45 45.8 197.0
+332829 NI 6000 0 10 12 71.6056 22.352 25.6 980.03 47.5 180.5
+332829 NI 6000 0 10 13 71.5783 22.5121 22.57 985.47 47.5 123.1
+332829 NI 6000 0 10 14 71.5359 22.6497 20.18 989.52 47.5 81.0
+332829 NI 6000 0 10 15 71.4936 22.7598 18.25 992.53 47.5 55.6
+332830 NI 6001 0 12 0 83.2017 15.5889 35.21 981.54 27.6 68.0
+332830 NI 6001 0 12 1 83.117 15.5165 33.76 983.16 30.8 81.4
+332830 NI 6001 0 12 2 82.8298 15.0915 32.43 984.51 34.1 94.9
+332830 NI 6001 0 12 3 82.5939 14.6754 31.2 985.86 37.3 108.4
+332830 NI 6001 0 12 4 82.408 14.2942 30.06 986.84 40.6 121.8
+332830 NI 6001 0 12 5 82.2396 13.8567 29.0 987.79 43.8 135.3
+332830 NI 6001 0 12 6 81.9361 13.3968 28.02 988.57 47.1 148.7
+332830 NI 6001 0 12 7 81.6539 12.9018 27.09 989.34 50.3 162.2
+332830 NI 6001 0 12 8 81.4142 12.3994 26.23 989.95 53.6 175.7
+332830 NI 6001 0 12 9 81.2358 11.9758 25.42 990.58 56.8 189.1
+332830 NI 6001 0 12 10 81.0424 11.5521 24.66 991.3 60.1 193.3
+332830 NI 6001 0 12 11 80.7916 11.1108 23.94 992.77 60.1 174.8
+332830 NI 6001 0 12 12 80.4705 10.6547 23.26 994.18 60.1 158.5
+332830 NI 6001 0 12 13 80.111 10.2192 22.62 995.42 60.1 144.1
+332830 NI 6001 0 12 14 79.6986 9.7073 22.02 996.57 60.1 131.4
+332830 NI 6001 0 12 15 79.2553 9.0699 21.44 997.41 60.1 120.1
+332830 NI 6001 0 12 16 78.9268 8.41 20.9 998.3 60.1 110.0
[…]
