### 2025-05-22 2733787
Add IRIS example TC tracks and function to load the tracks. This introduced a complication because the time is outside the usual np.datetime64 range when using nanosecond precision. Try to lower the time precision if this is the case, but this only works with very recent versions of xarray, so fall back on cftime if this does not work.
