+++ huracanpy/_data/_load.py
@@ -1,7 +1,14 @@
+import warnings
+from datetime import timedelta
+
+import cftime
+from dateutil.parser import parse
+
+import numpy as np
 import pandas as pd
+from pandas.errors import OutOfBoundsDatetime
 
-
-from . import _csv, _TRACK, _netcdf, _tempestextremes, witrack, _old_HURDAT
+from . import _csv, _TRACK, _netcdf, _tempestextremes, witrack, _old_HURDAT, iris_tc
 from . import ibtracs
 
 
@@ -16,7 +23,36 @@ rename_defaults = dict(
     stormID="track_id",
     # Names for TRACK netCDF
     TRACK_ID="track_id",
+    # Possible time names
+    iso_time="time",
+    isotime="time",
 )
+
+xarray_datetime_precisions = ["ns", "us", "ms", "s"]
+
+pandas_valid_time_labels = [
+    "year",
+    "years",
+    "month",
+    "months",
+    "day",
+    "days",
+    "hour",
+    "hours",
+    "minute",
+    "minutes",
+    "second",
+    "seconds",
+    "ms",
+    "millisecond",
+    "milliseconds",
+    "us",
+    "microsecond",
+    "microseconds",
+    "ns",
+    "nanosecond",
+    "nanoseconds",
+]
 
 
 def load(
@@ -28,6 +64,7 @@ def load(
     baselon=None,
     infer_track_id=None,
     ibtracs_subset="wmo",
+    iris_timestep=timedelta(hours=3),
     tempest_extremes_unstructured=False,
     tempest_extremes_header_str="start",
     track_calendar=None,
@@ -58,6 +95,7 @@ def load(
         * **csv**
         * **netcdf**, **nc** (includes support for CHAZ & MIT-Open file provided appropriate variable_names)
         * **old_hurdat**, **ecmwf**
+        * **iris**
 
     variable_names : list of str, optional
           When loading data from an ASCII file (TRACK or TempestExtremes), specify the
@@ -112,6 +150,12 @@ def load(
         * **since1980**: Entire IBTrACS database since 1980 (advent of satellite era,
           considered reliable from then on)
 
+    iris_timestep : int or datetime.timedelta, default=datetime.timedelta(hours=3)
+        The timestep used in the Imperial College Storm Model (IRIS). This is 3-hours
+        in the paper
+        (https://www.nature.com/articles/s41597-024-03250-y/tables/1), so unlikely to
+        need changing, but provided here in case it does.
+
     tempest_extremes_unstructured : bool, default=False
         By default the first two columns in TempestExtremes files are the i, j indices
         of the closest gridpoint, but for unstructured grids it is a single lookup index
@@ -161,11 +205,11 @@ def load(
     if source is None:
         extension = filename.split(".")[-1]
         if extension == "csv":
-            data = _csv.load(filename, **kwargs)
+            tracks = _csv.load(filename, **kwargs)
         elif extension == "parquet":
-            data = _csv.load(filename, load_function=pd.read_parquet, **kwargs)
+            tracks = _csv.load(filename, load_function=pd.read_parquet, **kwargs)
         elif filename.split(".")[-1] == "nc":
-            data = _netcdf.load(filename, rename, **kwargs)
+            tracks = _netcdf.load(filename, rename, **kwargs)
         else:
             raise ValueError("Source is set to None and file type is not detected")
 
@@ -173,53 +217,120 @@ def load(
     else:
         source = source.lower()
         if source == "track":
-            data = _TRACK.load(
+            tracks = _TRACK.load(
                 filename,
                 calendar=track_calendar,
                 variable_names=variable_names,
             )
         elif source == "track.tilt":
-            data = _TRACK.load_tilts(
[…]
