### 2025-05-22 5dcbe6f
Let xarray decide the time precision to use and force an error if it goes wrong
