### 2025-05-23 #8174 @malmans2
Hi @stella-bourdin,

I'm starting to review your submission today. Sorry for the delay!
The paper looks great overall. Just one quick comment: in the Loading Data section, are the * characters meant to be bullet points? If so, something seems off with the formatting.

Next, I'll review the software and I'll open issues in the GitHub repo if I have any suggestions.
