### 2025-05-23 Huracan-project/huracanpy#110 @malmans2
[JOSS-REVIEW] Improve Getting Started
Hi there,

I suggest adding another huracanpy-specific feature to the Getting Started section.
Something simple that your audience can easily pick up.

Right now, the page gives the impression that, apart from opening the format, everything else is standard xarray.
For example, you could at least mention and demonstrate that an accessor has been added to the xarray object.

https://github.com/openjournals/joss-reviews/issues/8174
