### 2025-05-23 #8174 @stella-bourdin
> in the Loading Data section, are the * characters meant to be bullet points?

Yes indeed! Thanks for catching it, I just fixed it.
