### 2025-05-23 Huracan-project/huracanpy#112 @malmans2
[JOSS-REVIEW] Improve Example Workflows
I think the Example Workflows look great, but I found that section a bit hard to follow.
Here are a few suggestions that might help improve clarity:
* Split each section into separate notebooks.
* Minimize the use of inline code comments, especially long ones. Instead, consider moving explanations to markdown cells.
* Follow Python PEP8 standards. For example, I'd recommend using lowercase names for variables such as `Wilma`, `ERA5`, and `LMI_wind_ib`, ...

https://github.com/openjournals/joss-reviews/issues/8174
