### 2025-05-23 Huracan-project/huracanpy#113 @malmans2
[JOSS-REVIEW] Missing links
It looks like there are a couple of missing links in the [Contributor's Guide](https://huracanpy.readthedocs.io/en/latest/dev_guide/index.html).
I see a few placeholder strings like `[LINK]`

https://github.com/openjournals/joss-reviews/issues/8174
