### 2025-05-23 Huracan-project/huracanpy#114 @malmans2
[JOSS-REVIEW] Rename master to main
I suggest renaming the master branch to main to align with current naming conventions: https://github.com/github/renaming?tab=readme-ov-f

https://github.com/openjournals/joss-reviews/issues/8174
