### 2025-05-23 Huracan-project/huracanpy#116 @malmans2
[JOSS-REVIEW] Improve code coverage
Here's the code coverage I'm seeing. Overall, it's not bad at all, but there are a few modules with lower coverage.
I just wanted to check whether you're already using a coverage tool and have considered improving coverage in those areas.
If you've already looked into it and further improvements aren't feasible, feel free to close this issue.

```
$ pip install pytest-cov
$ python -m pytest -vv --cov=. .
```

```
Name                                    Stmts   Miss  Cover
-----------------------------------------------------------
huracanpy/__init__.py                       7      0   100%
huracanpy/_accessor.py                    161     34    79%
huracanpy/_basins.py                       21      0   100%
huracanpy/_data/_TRACK.py                 120     15    88%
huracanpy/_data/__init__.py                19      0   100%
huracanpy/_data/_csv.py                    27      4    85%
huracanpy/_data/_load.py                   42      7    83%
huracanpy/_data/_netcdf.py                 60      1    98%
huracanpy/_data/_old_HURDAT.py             53      3    94%
huracanpy/_data/_save.py                    7      1    86%
huracanpy/_data/_tempestextremes.py        44      5    89%
huracanpy/_data/ibtracs.py                 25      5    80%
huracanpy/_data/witrack.py                 11      0   100%
huracanpy/_interp.py                       14      1    93%
huracanpy/_metpy.py                        15      0   100%
huracanpy/_subset.py                       16      2    88%
huracanpy/assess/__init__.py                4      0   100%
huracanpy/assess/_match.py                 42      6    86%
huracanpy/assess/_overlap.py               12      0   100%
huracanpy/assess/_scores.py                 9      0   100%
huracanpy/calc/__init__.py                  6      0   100%
huracanpy/calc/_density.py                 27      9    67%
huracanpy/calc/_lifecycle.py               17      3    82%
huracanpy/calc/_rates.py                   33      3    91%
huracanpy/calc/_track_stats.py             24      1    96%
huracanpy/calc/_translation.py             66      6    91%
huracanpy/info/__init__.py                  4      0   100%
huracanpy/info/_category.py                23      3    87%
huracanpy/info/_geography.py               46      0   100%
huracanpy/info/_time.py                    36      5    86%
huracanpy/plot/__init__.py                  6      0   100%
huracanpy/plot/_density.py                  8      5    38%
huracanpy/plot/_doughnut.py                16     14    12%
huracanpy/plot/_fancyline.py               47     39    17%
huracanpy/plot/_tracks.py                   8      4    50%
huracanpy/plot/_venn.py                    32     27    16%
huracanpy/tc/__init__.py                    4      0   100%
huracanpy/tc/_ace.py                       63     10    84%
huracanpy/tc/_beta_drift.py                28      4    86%
huracanpy/tc/_category.py                  16      0   100%
huracanpy/tc/_conventions.py                3      0   100%
setup.py                                    2      2     0%
tests/conftest.py                          37      2    95%
tests/test_accel.py                        99      0   100%
tests/test_accessor.py                     62      0   100%
tests/test_assess.py                       32      0   100%
tests/test_calc/test_lifecycle.py          11      0   100%
tests/test_calc/test_rates.py              34      0   100%
tests/test_calc/test_track_density.py       7      0   100%
tests/test_calc/test_track_stats.py        17      0   100%
tests/test_calc/test_translation.py        32      0   100%
tests/test_docs.py                         15      5    67%
tests/test_huracanpy.py                    52      0   100%
tests/test_info/test_category.py           13      0   100%
tests/test_info/test_geography.py          36      0   100%
tests/test_info/test_time.py               18      0   100%
tests/test_interpolate.py                   9      0   100%
tests/test_subset.py                       21      0   100%
tests/test_tc/test_ace.py                  19      0   100%
tests/test_tc/test_beta_drift.py            8      0   100%
tests/test_tc/test_tc_category.py          50      0   100%
-----------------------------------------------------------
TOTAL                                    1796    226    87%
```

https://github.com/openjournals/joss-reviews/issues/8174#issuecomment-2904664255
