### 2025-05-26 Huracan-project/huracanpy#117 @malmans2
[JOSS-REVIEW] Add a clearly-stated list of dependencies
I couldn't find a section that clearly lists the dependencies.
For reference, here's how `xarray` exposes this in their documentation: https://docs.xarray.dev/en/stable/getting-started-guide/installing.html

https://github.com/openjournals/joss-reviews/issues/8174
