+++ .github/workflows/python-package.yaml
@@ -5,9 +5,9 @@ name: Python package
 
 on:
   push:
-    branches: [ "master", "v1.2" ]
+    branches: [ "main", "v1.2" ]
   pull_request:
-    branches: [ "master", "v1.2" ]
+    branches: [ "main", "v1.2" ]
 
 jobs:
   build:
