### 2025-05-27 f52abef
Generate coverage html report by default when running pytest
