### 2025-05-27 Huracan-project/huracanpy#114 @leosaffin
Thanks for the prompt on this one. It's something I had wanted to do and assumed was going to be much more difficult that it actually was.
