### 2025-05-27 c205d15
Increase test coverage where possible and fix a few errors. 1. Consistent use of dummy track_id in rates/translation when not specified 2. Correctly convert units for category, when they are only included for one variable 3. Always convert units to metres in beta_drift
