+++ huracanpy/info/_time.py
@@ -43,10 +43,13 @@ def season(track_id, lat, time, convention="tc-short"):
     lat : xarray.DataArray
     time : xarray.DataArray
     convention : str
-        * 'tc-short' : In the Northern hemisphere, the season is the same as calendar year.
-        In the southern hemisphere, the season n corresponds to July n-1 to June n
-        * 'tc-long' : In the Northern hemisphere, the season is the same as calendar year.
-        In the southern hemisphere, the season from July n-1 to June n is named "(n-1)n"
+
+        * 'tc-short' : In the Northern hemisphere, the season is the same as calendar
+          year. In the southern hemisphere, the season n corresponds to July n-1 to
+          June n
+        * 'tc-long' : In the Northern hemisphere, the season is the same as calendar
+          year. In the southern hemisphere, the season from July n-1 to June n is named
+          "(n-1)n"
 
     Raises
     ------
@@ -56,8 +59,8 @@ def season(track_id, lat, time, convention="tc-short"):
     Returns
     -------
     xarray.DataArray
-        The season series.
-        You can append it to your tracks by running tracks["season"] = get_season(tracks.track_id, tracks.lat, tracks.time)
+        The season series. You can append it to your tracks by running
+        tracks["season"] = get_season(tracks.track_id, tracks.lat, tracks.time)
     """
 
     # Derive values
