### 2025-05-28 5b0cd7b
Add test for missing accessor functions and implement currently missing ones (beta_drift, fancyline)
