### 2025-05-28 2894cb0
Test coverage for all non-plot accessor functions. Fixes 1. Consistent use of dummy track id in calc delta style functions 2. Allow for different time components in add_time_components 3. Consistent use of var_name 4. Fix call to azimuth
