### 2025-05-28 b7ec53c
multiply kde so that integral is number of points
