### 2025-05-28 9dafb60
Test coverage for plots and accessor versions. Fixes 1. Allow plot.density and plot.tracks to be used with existing axes 2. Fix filling of extra_variables for accessor version of fancyline
