+++ .github/workflows/python-package.yaml
@@ -25,13 +25,13 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
-        conda install -c conda-forge cartopy
+        python -m conda install -c conda-forge cartopy
         python -m pip install --upgrade pip
-        pip install .[dev]
+        python -m pip install .[dev]
     - uses: chartboost/ruff-action@v1
     - uses: chartboost/ruff-action@v1
       with:
         args: 'format --check'
     - name: Test with pytest
       run: |
-        pytest
+        python -m pytest
