+++ .github/workflows/python-package.yaml
@@ -22,6 +22,7 @@ jobs:
     - uses: actions/checkout@v4
     - uses: mamba-org/setup-micromamba@v2
       with:
+        environment-name: test-huracanpy
         create-args: >-
           python=${{matrix.python-version}}
     - name: Install dependencies
