### 2025-05-29 72c71ce
Try explicitly activating the new environment
