+++ .github/workflows/python-package.yaml
@@ -27,7 +27,8 @@ jobs:
           python=${{matrix.python-version}}
     - name: Install dependencies
       run: |
-        mamba activate test-huracanpy
+        python -c "import sys; print(sys.version)"
+        micromamba activate test-huracanpy
         python -c "import sys; print(sys.version)"
         python -m pip install .[dev]
     - uses: chartboost/ruff-action@v1
