### 2025-05-29 0bd9680
Seems to have worked so just run pytest in the same section
