### 2025-05-29 2e18019
Implement area-proportional kde track density
