### 2025-06-02 0cca403
Replace uses of np.array with np.asarray to remove deprecation warning
