+++ tests/test_plot/test_doughnut.py
@@ -17,24 +17,24 @@ def test_doughnut_roberts2020():
     # A 2d array of average number of storms per year as a function of model and basin
     da = xr.DataArray(
         data=[
-            np.array([16, 53, 19, 5, 7]) * 0.521,
-            np.array([17, 46, 23, 7, 7]) * 0.873,
-            np.array([16, 46, 16, 17, 5]) * 0.509,
-            np.array([17, 46, 18, 15, 5]) * 0.581,
-            np.array([13, 45, 20, 18, 4]) * 0.249,
-            np.array([14, 44, 20, 18, 4]) * 0.42,
-            np.array([22, 42, 12, 18, 5]) * 0.662,
-            np.array([23, 44, 9, 20, 4]) * 0.654,
-            np.array([21, 33, 19, 22, 5]) * 0.14,
-            np.array([19, 35, 19, 22, 4]) * 0.133,
-            np.array([11, 32, 21, 28, 8]) * 0.296,
-            np.array([14, 39, 21, 18, 8]) * 0.649,
-            np.array([17, 40, 28, 11, 4]) * 0.648,
-            np.array([19, 39, 26, 10, 6]) * 0.627,
-            np.array([18, 45, 20, 12, 5]) * 0.495,
-            np.array([18, 42, 23, 13, 4]) * 0.617,
-            np.array([21, 38, 24, 11, 6]) * 0.729,
-            np.array([21, 42, 28, 8, 1]) * 0.536,
+            np.asarray([16, 53, 19, 5, 7]) * 0.521,
+            np.asarray([17, 46, 23, 7, 7]) * 0.873,
+            np.asarray([16, 46, 16, 17, 5]) * 0.509,
+            np.asarray([17, 46, 18, 15, 5]) * 0.581,
+            np.asarray([13, 45, 20, 18, 4]) * 0.249,
+            np.asarray([14, 44, 20, 18, 4]) * 0.42,
+            np.asarray([22, 42, 12, 18, 5]) * 0.662,
+            np.asarray([23, 44, 9, 20, 4]) * 0.654,
+            np.asarray([21, 33, 19, 22, 5]) * 0.14,
+            np.asarray([19, 35, 19, 22, 4]) * 0.133,
+            np.asarray([11, 32, 21, 28, 8]) * 0.296,
+            np.asarray([14, 39, 21, 18, 8]) * 0.649,
+            np.asarray([17, 40, 28, 11, 4]) * 0.648,
+            np.asarray([19, 39, 26, 10, 6]) * 0.627,
+            np.asarray([18, 45, 20, 12, 5]) * 0.495,
+            np.asarray([18, 42, 23, 13, 4]) * 0.617,
+            np.asarray([21, 38, 24, 11, 6]) * 0.729,
+            np.asarray([21, 42, 28, 8, 1]) * 0.536,
         ],
         coords=dict(
             model=[
