+++ tests/test_subset.py
@@ -20,7 +20,7 @@ def test_sel_id_fails(tracks_csv):
     with pytest.raises(ValueError, match="track_ids must be 1d"):
         huracanpy.sel_id(
             tracks_csv,
-            np.array([tracks_csv.track_id.values, tracks_csv.track_id.values]),
+            np.asarray([tracks_csv.track_id.values, tracks_csv.track_id.values]),
             0,
         )
 
@@ -51,6 +51,6 @@ def test_trackswhere_fails(tracks_csv):
     with pytest.raises(ValueError, match="track_ids must be 1d"):
         huracanpy.trackswhere(
             tracks_csv,
-            np.array([tracks_csv.track_id.values, tracks_csv.track_id.values]),
+            np.asarray([tracks_csv.track_id.values, tracks_csv.track_id.values]),
             lambda track: (track.track_id == 0).all(),
         )
