### 2025-06-02 06970b0
Unify unit validation
