+++ huracanpy/_metpy.py
@@ -1,7 +1,8 @@
 import functools
 
-import xarray as xr
+from metpy.units import units
 import pint
+import xarray as xr
 
 
 def dequantify_results(original_function):
@@ -20,5 +21,16 @@ def dequantify_results(original_function):
 def _dequantify_result(result):
     if isinstance(result, xr.DataArray) and isinstance(result.data, pint.Quantity):
         return result.metpy.dequantify()
+    elif isinstance(result, pint.Quantity) and result.unitless:
+        return result.magnitude
     else:
         return result
+
+
+def validate_units(variable, expected_units):
+    if not isinstance(variable, pint.Quantity) or variable.unitless:
+        if callable(expected_units):
+            expected_units = expected_units(variable)
+        variable = variable * units(expected_units)
+
+    return variable
