+++ tests/test_tc/test_ace.py
@@ -1,5 +1,6 @@
 from metpy.units import units
 import numpy as np
+from numpy.polynomial.polynomial import Polynomial
 import pytest
 
 import huracanpy
@@ -29,12 +30,33 @@ def test_ace(tracks_csv, threshold, sum_by, result):
 
 
 @pytest.mark.parametrize(
-    "threshold_pressure, threshold_wind, sum_by, result",
+    "model, threshold_pressure, threshold_wind, sum_by, kwargs, result",
     [
-        (None, None, "track_id", [4.34978137, 2.65410482, 6.09892875]),
+        (None, None, None, "track_id", {}, [4.34978137, 2.65410482, 6.09892875]),
+        (
+            Polynomial,
+            None,
+            None,
+            "track_id",
+            {"deg": 2},
+            [4.34978137, 2.65410482, 6.09892875],
+        ),
+        (None, 1e5, None, "track_id", {}, [4.115545, 2.654105, 5.152805]),
+        (
+            None,
+            None,
+            17 * units("m s-1"),
+            "track_id",
+            {},
+            [3.332754, 2.281562, 4.49114],
+        ),
+        ("z2021", None, None, "track_id", {}, [2.266398, 1.42874, 3.19211]),
+        ("holland", None, None, "track_id", {}, [2.03433554, 1.23160966, 2.72186645]),
     ],
 )
-def test_pace(tracks_csv, threshold_pressure, threshold_wind, sum_by, result):
+def test_pace(
+    tracks_csv, model, threshold_pressure, threshold_wind, sum_by, kwargs, result
+):
     if sum_by is not None:
         sum_by = tracks_csv[sum_by]
 
@@ -42,9 +64,12 @@ def test_pace(tracks_csv, threshold_pressure, threshold_wind, sum_by, result):
     pace, model = huracanpy.tc.pace(
         tracks_csv.slp,
         tracks_csv.wind10,
+        model=model,
         sum_by=sum_by,
         threshold_wind=threshold_wind,
         threshold_pressure=threshold_pressure,
+        pressure_units="Pa",
+        **kwargs,
     )
 
     np.testing.assert_allclose(pace, result)
@@ -56,6 +81,13 @@ def test_pace(tracks_csv, threshold_pressure, threshold_wind, sum_by, result):
         sum_by=sum_by,
         threshold_wind=threshold_wind,
         threshold_pressure=threshold_pressure,
+        pressure_units="Pa",
+        **kwargs,
     )
 
-    np.testing.assert_allclose(pace, np.array([4.34978137, 2.65410482, 6.09892875]))
+    np.testing.assert_allclose(pace, result)
+
+
+def test_pace_fails(tracks_csv):
+    with pytest.raises(ValueError, match="Need to specify either wind or model"):
+        huracanpy.tc.pace(tracks_csv.slp)
