### 2025-06-02 d0ce6b8
Ignore ruff check on lambda function
