### 2025-06-02 58a0416
Improve load test coverage. Further coverage needs fake files with bad formatting
