### 2025-06-02 f37bbbd
Fix error messages when building docs
