+++ docs/conf.py
@@ -57,8 +57,6 @@ html_theme = "pydata_sphinx_theme"
 html_theme_options = {
     "github_url": "https://github.com/Huracan-project/huracanpy",
     "show_toc_level": 3,
-    "logo_only": True,
-    "display_version": False,
 }
 html_static_path = ["_static"]
 html_logo = "images/logo/logo-with-name/Slide1.png"
