### 2025-06-02 Huracan-project/huracanpy#113 @leosaffin
Fixed with #125
