### 2025-06-03 Huracan-project/huracanpy#117 @leosaffin
Added a list of dependencies with #126
