### 2025-06-03 a312ed7
Add stricter ruff line length rules and correct errors
