+++ huracanpy/plot/_density.py
@@ -28,7 +28,7 @@ def density(
     contourf_kws : dict, optional
         Arguments to be passed to :func:`matplotlib.pyplot.contourf`
     subplot_kws : dict, optional
-        Arguments to be passed to :func:`matplotlib.pyplot.subplots`' subplot_kw argument
+        Arguments to be passed to :func:`matplotlib.pyplot.subplots` subplot_kw argument
     fig_kws : dict, optional
         Arguments to be passed to :func:`matplotlib.pyplot.subplots`
     cbar_kwargs: dict, optional
