+++ pyproject.toml
@@ -65,6 +65,7 @@ extend-select = [
     "ARG", # Flake8 unused arguments
     "S", # Flake8 bandit
     "T20", # Flake8 print
+    "E501", # Line too long
 ]
 
 [tool.ruff.lint.per-file-ignores]
