### 2025-06-03 1c592e3
Add pep8 naming conventions to ruff and fix errors. Note that this did not affect ipynb's because the lower case naming is only checked for variables within a function
