+++ tests/test_calc/test_track_stats.py
@@ -13,16 +13,16 @@ def test_duration():
 
 def test_gen_vals():
     data = huracanpy.load(huracanpy.example_csv_file, source="csv")
-    G = huracanpy.calc.gen_vals(data, data.time, data.track_id)
-    assert G.time.dt.day.mean() == 10
+    g = huracanpy.calc.gen_vals(data, data.time, data.track_id)
+    assert g.time.dt.day.mean() == 10
 
 
 def test_apex_vals():
     data = huracanpy.load(huracanpy.example_csv_file)
-    M = huracanpy.calc.apex_vals(data, data.wind10, data.track_id, "max")
-    m = huracanpy.calc.apex_vals(data, data.slp, data.track_id, "min")
-    assert M.time.dt.day.mean() == 15
-    assert m.lat.mean() == -27
+    max_wind = huracanpy.calc.apex_vals(data, data.wind10, data.track_id, "max")
+    min_pressure = huracanpy.calc.apex_vals(data, data.slp, data.track_id, "min")
+    assert max_wind.time.dt.day.mean() == 15
+    assert min_pressure.lat.mean() == -27
 
 
 def test_apex_vals_fails(tracks_csv):
