+++ tests/test_calc/test_track_density.py
@@ -8,10 +8,10 @@ import huracanpy
 @pytest.mark.parametrize("crop", [True, False])
 def test_density(baselon, method, crop):
     data = huracanpy.load(huracanpy.example_year_file, baselon=baselon)
-    D = huracanpy.calc.density(data.lon, data.lat, method=method, crop=crop)
-    assert D.min() == 0.0
-    assert D.max() == 43.0
-    assert D.sum() == len(data.record)
+    d = huracanpy.calc.density(data.lon, data.lat, method=method, crop=crop)
+    assert d.min() == 0.0
+    assert d.max() == 43.0
+    assert d.sum() == len(data.record)
 
 
 def test_track_density_fails():
