+++ pyproject.toml
@@ -66,6 +66,7 @@ extend-select = [
     "S", # Flake8 bandit
     "T20", # Flake8 print
     "E501", # Line too long
+    "N", # pep8 naming conventions
 ]
 
 [tool.ruff.lint.per-file-ignores]
