### 2025-06-04 cbcd1d1
Split demo into 3 separate notebooks
