### 2025-06-04 Huracan-project/huracanpy#112 @leosaffin
Added the suggested improvements with #128. Thanks
