### 2025-06-04 6472507
Add load_list function for loading and concatenating several files at once
