### 2025-06-04 776a140
Add a test for load_list
