### 2025-06-04 d602e5a
Add docstring to load_list
