### 2025-06-05 0ec7b93
Add test for STORM sample file with no 'day' nor 'track_id' data
