### 2025-06-05 #8174 @malmans2
Hi @hugoledoux,

The authors have addressed all the suggestions I made during the review. I find `huracanpy` to be a useful and well-documented package that brings clear value to the scientific community. It meets the standards of JOSS, and I recommend it for publication.
