### 2025-06-05 fde9970
Fix unit warning filter for season
