### 2025-06-05 3919dcc
Add load_list function
