+++ huracanpy/_data/_load.py
@@ -324,3 +324,31 @@ def load(
         tracks = tracks.hrcn.add_inferred_track_id(*infer_track_id)
 
     return tracks
+
+def load_list(filelist, **kwargs):
+    """
+    This function opens all the files in a list and concatenate them. All files should be opened with the exact same load command.
+    Track ids will be made unique by appending an index at the start.
+
+    Parameters
+    ----------
+    filename : list or np.ndarray
+        The list of file to be opened
+        
+    kwargs: 
+        Any parameter you would give to huracanpy.load to load individual files
+
+    Returns
+    -------
+    xarray.Dataset
+    """
+    # Loop through all the files and open them
+    tracks = []
+    for i, filepath in enumerate(tqdm(filelist)):
+        data = load(filepath, **kwargs)
+        if "tracks" in data.dims:
+            data = data.drop_dims("tracks")
+        data["track_id"] = str(i) + '-' + data["track_id"].astype(str) # Make sure track_ids remain unique
+        tracks.append(data)
+    # Concatenate in one object
+    return xr.concat(tracks, dim = "record")
