### 2025-06-05 58a8acc
Add test for load list
