+++ huracanpy/_data/_load.py
@@ -1,11 +1,13 @@
 import warnings
 from datetime import timedelta
+from tqdm import tqdm
 
 import cftime
 from dateutil.parser import parse
 
 import numpy as np
 import pandas as pd
+import xarray as xr
 from pandas.errors import OutOfBoundsDatetime
 
 from . import _csv, _TRACK, _netcdf, _tempestextremes, witrack, _old_HURDAT, iris_tc
@@ -325,6 +327,7 @@ def load(
 
     return tracks
 
+
 def load_list(filelist, **kwargs):
     """
     This function opens all the files in a list and concatenate them. All files should be opened with the exact same load command.
@@ -334,8 +337,8 @@ def load_list(filelist, **kwargs):
     ----------
     filename : list or np.ndarray
         The list of file to be opened
-        
-    kwargs: 
+
+    kwargs:
         Any parameter you would give to huracanpy.load to load individual files
 
     Returns
@@ -348,7 +351,9 @@ def load_list(filelist, **kwargs):
         data = load(filepath, **kwargs)
         if "tracks" in data.dims:
             data = data.drop_dims("tracks")
-        data["track_id"] = str(i) + '-' + data["track_id"].astype(str) # Make sure track_ids remain unique
+        data["track_id"] = (
+            str(i) + "-" + data["track_id"].astype(str)
+        )  # Make sure track_ids remain unique
         tracks.append(data)
     # Concatenate in one object
-    return xr.concat(tracks, dim = "record")
+    return xr.concat(tracks, dim="record")
