### 2025-06-05 9ef417b
Remember basic trigonometry
