### 2025-06-05 Huracan-project/huracanpy#104 @nmstreethran
Thanks for addressing my comments and sorry for the delay in getting back to you. Just a minor comment on the command for installing directly from the repository. It only works for me when the package name and path are enclosed in quotes as follows:

```sh
python -m pip install "huracanpy @ git+https://github.com/Huracan-project/huracanpy"
```

Might be worth testing on your side and updating the docs and README.
