### 2025-06-05 Huracan-project/huracanpy#132 @nmstreethran
[JOSS Review] Paper comments
Some comments on your JOSS paper (https://github.com/openjournals/joss-reviews/issues/8174)

- "Python", "Xarray", "MetPy", and "Pint" should be capitalised.
- There are a couple of typos:
  - "Manipulating" in the first bullet point under Analysis.
  - "developed" in the second paragraph of the Acknowledgements.
- JOSS requires the Statement of Need (in the paper and the docs) to include the who the target audience is. Could you add this?
- The second sentence of the Summary states "HuracanPy provides a powerful loading function that supports **most of the most common** track data ...". The part in bold could be rephrased a little.
- Under Loading data:
  - CSV is defined as comma-separated variable. I believe this should be comma-separated values.
  - GFDL is not defined.
  - Hodges (1995) and (1999) are in the References list but only the latter is cited in the text.
- There is no section on the state of the field describing how this software compares to other commonly-used packages. Can this be added?
- Under Perspectives, "TC" in "TC-specific diagnoses" is not defined.
- Under Perspectives and Acknowledgements, the final sentences are missing periods at the end.
