### 2025-06-06 5c6995c
Fix install instructions for git install
