### 2025-06-06 Huracan-project/huracanpy#104 @leosaffin
Thanks for checking that. I've updated and tested that with #133
