### 2025-06-06 e01a956
Fix broken inferred track_id test from merge
