### 2025-06-06 e9da166
Implement reviewer's comments
