+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'HuracanPy: A python package for reading and analysing cyclone tracks'
+title: 'HuracanPy: A Python package for reading and analysing cyclone tracks'
 tags:
   - Python
   - cyclones
@@ -28,10 +28,10 @@ bibliography: paper.bib
 
 # Summary
 
-HuracanPy is a python module for working with cyclone track data, where cyclone tracks are recorded as a list of points in space and time. 
-HuracanPy provides a powerful loading function that supports most of the most common track data formats and loads the tracks as xarray datasets [@xarray2017].
-As such, we draw from the power of xarray's built-in functions and the functionality of other libraries that work with xarray, such as MetPy [@metpy]. 
-We also provide functions specific to cyclone analysis, which are gathered as methods in an xarray accessor. 
+HuracanPy is a Python module for working with cyclone track data, where cyclone tracks are recorded as a list of points in space and time. 
+HuracanPy provides a powerful loading function that supports all the track data formats the authors are aware of, and loads the tracks as Xarray datasets [@xarray2017].
+As such, we draw from the power of Xarray's built-in functions and the functionality of other libraries that work with Xarray, such as MetPy [@metpy]. 
+We also provide functions specific to cyclone analysis, which are gathered as methods in an Xarray accessor. 
 These include functions to subset or interpolate track data, to add useful information, to compute common diagnoses, and to make quick exploration plots. 
 With this package, we hope to make track data analysis more accessible and more reproducible. 
 
@@ -57,18 +57,18 @@ Our aim is to make track data analysis more accessible, and to promote good repr
 
 # Description
 
-HuracanPy is built on xarray. It loads track data as an xarray dataset, and provides an xarray accessor to provide cyclone-specific analysis methods. 
-This allows us to draw from xarray's performance and existing methods, and also means that we can handle track attributes which are not just scalar.
+HuracanPy is built on Xarray. It loads track data as an Xarray dataset, and provides an Xarray accessor to provide cyclone-specific analysis methods. 
+This allows us to draw from Xarray's performance and existing methods, and also means that we can handle track attributes which are not just scalar.
 
 ## Loading data
 
 HuracanPy currently supports loading tracks from:
 
-* CSV (comma-separated variable) files where each row correspond to a cyclone point and each column to a feature.
+* CSV (comma-separated values) files where each row corresponds to a cyclone point and each column to a feature.
   
 * NetCDF (Network Common Data Form) files following CF (Climate and Forecast) conventions [@cfconventions], specifically the formats for trajectory data described in H.4 [@cfconventionswebpage].
 
-* NetCDF files similar to the CF conventions can also be loaded provided they use specific naming for the track ID.
+* NetCDF files similar to the CF conventions can also be loaded, provided they use specific naming for the track ID.
 
 * Text files from TempestExtreme's StitchNodes (GFDL format, see [@ullrich2021tempestextremes])
 
@@ -76,16 +76,16 @@ HuracanPy currently supports loading tracks from:
 
 HuracanPy can also load IBTrACS [@gahtan2024international] from the online source, or from an embedded file. 
 
-Tracks are loaded as an xarray dataset with one main dimension called `record`. 
+Tracks are loaded as an Xarray dataset with one main dimension called `record`. 
 
 HuracanPy can also save the track data as CSV and NetCDF files. 
 
 ## Analysis
 
-Most of HuracanPy's analysis functions can be called through an xarray accessor named `hrcn`.
+Most of HuracanPy's analysis functions can be called through an Xarray accessor named `hrcn`.
 
-* Manupulating data:
-    Loading the track data as an xarray object means the user can easily use xarray's methods such as `.where` to subset. 
[…]
