### 2025-06-06 e02481b
Update paper.md
