+++ paper/paper.bib
@@ -144,3 +144,14 @@ Publisher: American Meteorological Society},
   author={Gahtan, J and Knapp, KR and Schreck, CJ and Diamond, HJ and Kossin, JP and Kruk, MC},
   year={2024}
 }
+
+@article{sparks2024imperial,
+  title={The ImpeRIal college storm model (IRIS) dataset},
+  author={Sparks, Nathan and Toumi, Ralf},
+  journal={Scientific data},
+  volume={11},
+  number={1},
+  pages={424},
+  year={2024},
+  publisher={Nature Publishing Group UK London}
+}
