### 2025-06-06 #8174 @stella-bourdin
Implementing comments by @nmstreethran in issue https://github.com/Huracan-project/huracanpy/issues/132
Except one pending about comparing with other packages.
