### 2025-06-06 8a8d911
Put load_list functionality into load. Add two new functions, concat_tracks and reset_track_id to use for this
