+++ huracanpy/_accessor.py
@@ -2,8 +2,17 @@ import xarray as xr
 from metpy.units import units
 import pandas as pd
 
-import huracanpy
-from . import plot, tc, info, calc, save, interp_time, sel_id, trackswhere
+from . import (
+    plot,
+    tc,
+    info,
+    calc,
+    save,
+    interp_time,
+    sel_id,
+    trackswhere,
+    reset_track_id,
+)
 
 
 @xr.register_dataarray_accessor("hrcn")
@@ -53,6 +62,14 @@ class HuracanPyDatasetAccessor:
 
     def trackswhere(self, condition, track_id_name="track_id"):
         return trackswhere(self._dataset, self._dataset[track_id_name], condition)
+
+    def reset_track_id(self, track_id_name="track_id", start=0, keep_original=False):
+        return reset_track_id(
+            self._dataset,
+            self._dataset[track_id_name],
+            start=start,
+            keep_original=keep_original,
+        )
 
     # %% utils
     # ---- geography
@@ -259,9 +276,7 @@ class HuracanPyDatasetAccessor:
 
     # --- utils
     def get_inferred_track_id(self, *variable_names):
-        return huracanpy.info.inferred_track_id(
-            *[self._dataset[var] for var in variable_names]
-        )
+        return info.inferred_track_id(*[self._dataset[var] for var in variable_names])
 
     def add_inferred_track_id(self, *variable_names, track_id_name="track_id"):
         self._dataset[track_id_name] = self.get_inferred_track_id(*variable_names)
