+++ tests/test_accessor.py
@@ -12,6 +12,7 @@ import huracanpy
 # Functions not in the accessor
 _intentionally_missing = [
     "load",
+    "concat_tracks",
     # add_ functions that would have output with a different shape
     "add_apex_vals",
     "add_gen_vals",
