+++ docs/api/huracanpy.rst
@@ -12,3 +12,5 @@ huracanpy
     sel_id
     trackswhere
     interp_time
+    concat_tracks
+    reset_track_id

[…]
