### 2025-06-09 de453a6
Move concat_tracks and hide reset_track_id
