### 2025-06-09 b5ff796
Ruff
