+++ paper/paper.bib
@@ -161,3 +161,13 @@ Publisher: American Meteorological Society},
   year={2024},
   publisher={Nature Publishing Group UK London}
 }
+
+@article{zarzycki2021metrics,
+  title={Metrics for evaluating tropical cyclones in climate data},
+  author={Zarzycki, Colin M and Ullrich, Paul A and Reed, Kevin A},
+  journal={Journal of Applied Meteorology and Climatology},
+  volume={60},
+  number={5},
+  pages={643--660},
+  year={2021}
+}
