### 2025-06-12 a68e8ce
Update paper.md
