### 2025-06-12 #8174 @stella-bourdin
We have added a comparison to other existing packages
