### 2025-06-13 Huracan-project/huracanpy#132 @nmstreethran
Thank you for implementing the changes. I've checked the PDF of the paper and have just a few minor comments:

- Under Comparison with other packages, HuracanPy is misspelled as "HruacanPy"
- I think you meant to add bullet points under the comparison but it did not render properly
- Under Loading data, the reference [sparks2024imperial] has a syntax issue

All good otherwise.
