### 2025-06-19 #8174 @hugoledoux
okay, thanks everyone for the swift reviews.

I have myself read the paper and fixes many small typos and errors, see: https://github.com/Huracan-project/huracanpy/pull/140

I think it's good if you go over the whole paper one more time. Then I can move on with the acceptance
