### 2025-06-20 #8174 @hugoledoux
@editoralbot check references
