+++ paper/paper.bib
@@ -57,7 +57,7 @@
   author = {Ryan M. May and Kevin H. Goebbert and Jonathan E. Thielen and John R. Leeman and
             M. Drew Camron and Zachary Bruick and Eric C. Bruning and Russell P. Manser and
             Sean C. Arms and Patrick T. Marsh},
-  title = {MetPy: A Meteorological Python Library for Data Analysis and Visualization},
+  title = {MetPy: A Meteorological {Python} Library for Data Analysis and Visualization},
   journal = {Bulletin of the American Meteorological Society},
   year = {2022},
   publisher = {American Meteorological Society},
@@ -146,7 +146,7 @@ Publisher: American Meteorological Society},
 }
 
 @misc{gahtan2024international,
-  title={International Best Track Archive for Climate Stewardship (IBTrACS) Project, Version 4r01, NOAA National Centers for Environmental Information},
+  title={International Best Track Archive for Climate Stewardship ({IBTrACS}) Project, Version 4r01, {NOAA} National Centers for Environmental Information},
   author={Gahtan, J and Knapp, KR and Schreck, CJ and Diamond, HJ and Kossin, JP and Kruk, MC},
   year={2024}
 }
