### 2025-06-20 #8174 @stella-bourdin
I also proof-read the paper and fixed a few more typos.
