### 2025-06-20 #8174 @hugoledoux
> Is it ok with the pull request I just merged?

Yes it was, sorry wrong URL...
