### 2025-06-26 #8174 @leosaffin
@kthyng the capitalisation looks good in the most recent proof, but I see in out .bib file we have some titles just wrapped once with "{}", and others with specific words/letters wrapped. Are they OK with the single wrapped since it looks fine in the proof, or do we need to add extra wrappers?
