### 2025-06-27 #8174 @leosaffin
@kthyng Thanks. I had missed that and a few others. I got confused looking at the mixed case words which didn't need wrapping. Updated the branch now
